# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Rest interactiva y pública para el proyecto Anirank.",
        "title": "Anirank API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Soporte API Anirank",
            "email": "admin@anirank.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/tournaments/active": {
            "get": {
                "description": "Fetch the active tournament and its current matchups/bracket.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tournaments"
                ],
                "summary": "Get Active Tournament",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "data": {
                                    "$ref": "#/definitions/domain.Tournament"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/tournaments/matchups/{id}/vote": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Submit a vote for a specific song in an ongoing tournament matchup.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tournaments"
                ],
                "summary": "Submit Vote",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Matchup ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Vote Data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "song_id": {
                                    "type": "integer"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/domain.AppError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/domain.AppError"
                        }
                    }
                }
            }
        },
        "/tournaments/{slug}": {
            "get": {
                "description": "Fetch a specific tournament's details and matchups by its slug.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tournaments"
                ],
                "summary": "Get Tournament by Slug",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tournament Slug",
                        "name": "slug",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "data": {
                                    "$ref": "#/definitions/domain.Tournament"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/domain.AppError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "domain.AppError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "err": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "domain.Artist": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "name_jp": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "domain.ExternalLink": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "icon": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "domain.Format": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "domain.Genre": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "domain.Post": {
            "type": "object",
            "properties": {
                "anilist_id": {
                    "type": "integer"
                },
                "banner": {
                    "type": "string"
                },
                "banner_src": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "external_links": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.ExternalLink"
                    }
                },
                "format": {
                    "$ref": "#/definitions/domain.Format"
                },
                "format_id": {
                    "type": "integer"
                },
                "genres": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.Genre"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "producers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.Producer"
                    }
                },
                "season": {
                    "$ref": "#/definitions/domain.Season"
                },
                "season_id": {
                    "type": "integer"
                },
                "slug": {
                    "type": "string"
                },
                "songs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.Song"
                    }
                },
                "status": {
                    "type": "boolean"
                },
                "studios": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.Studio"
                    }
                },
                "thumbnail": {
                    "type": "string"
                },
                "thumbnail_src": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "year": {
                    "description": "Relations",
                    "allOf": [
                        {
                            "$ref": "#/definitions/domain.Year"
                        }
                    ]
                },
                "year_id": {
                    "type": "integer"
                }
            }
        },
        "domain.Producer": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "domain.Season": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "current": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "domain.Song": {
            "type": "object",
            "properties": {
                "anime_id": {
                    "type": "integer"
                },
                "artists": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.Artist"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "post": {
                    "description": "Relations",
                    "allOf": [
                        {
                            "$ref": "#/definitions/domain.Post"
                        }
                    ]
                },
                "season": {
                    "$ref": "#/definitions/domain.Season"
                },
                "season_id": {
                    "type": "integer"
                },
                "slug": {
                    "type": "string"
                },
                "song_en": {
                    "type": "string"
                },
                "song_jp": {
                    "type": "string"
                },
                "song_romaji": {
                    "type": "string"
                },
                "theme_num": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "variants": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.SongVariant"
                    }
                },
                "views": {
                    "type": "integer"
                },
                "year": {
                    "$ref": "#/definitions/domain.Year"
                },
                "year_id": {
                    "type": "integer"
                }
            }
        },
        "domain.SongVariant": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "embed_code": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "season": {
                    "$ref": "#/definitions/domain.Season"
                },
                "season_id": {
                    "type": "integer"
                },
                "slug": {
                    "type": "string"
                },
                "song": {
                    "description": "Relations",
                    "allOf": [
                        {
                            "$ref": "#/definitions/domain.Song"
                        }
                    ]
                },
                "song_id": {
                    "type": "integer"
                },
                "spoiler": {
                    "type": "boolean"
                },
                "updated_at": {
                    "type": "string"
                },
                "version_number": {
                    "type": "integer"
                },
                "video_src": {
                    "description": "Embedded Video Info to simulate 1:1 legacy relation without join overhead",
                    "type": "string"
                },
                "video_type": {
                    "description": "file, embed",
                    "type": "string"
                },
                "views": {
                    "type": "integer"
                },
                "year": {
                    "$ref": "#/definitions/domain.Year"
                },
                "year_id": {
                    "type": "integer"
                }
            }
        },
        "domain.Studio": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "domain.Tournament": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "current_round": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "matchups": {
                    "description": "Tree of matchups",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.TournamentMatchup"
                    }
                },
                "name": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "status": {
                    "description": "draft, active, completed",
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "domain.TournamentMatchup": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "matchup_order": {
                    "type": "integer"
                },
                "round": {
                    "type": "integer"
                },
                "song_a": {
                    "description": "Relations",
                    "allOf": [
                        {
                            "$ref": "#/definitions/domain.Song"
                        }
                    ]
                },
                "song_a_id": {
                    "type": "integer"
                },
                "song_b": {
                    "$ref": "#/definitions/domain.Song"
                },
                "song_b_id": {
                    "description": "Nulable if waiting or got a bye",
                    "type": "integer"
                },
                "status": {
                    "description": "pending, active, completed",
                    "type": "string"
                },
                "tournament_id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "winner": {
                    "$ref": "#/definitions/domain.Song"
                },
                "winner_id": {
                    "type": "integer"
                }
            }
        },
...
        "domain.Year": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "current": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}